refactor(sps): consolidate 5 source traits into one BackfillSource
Following the discussion in PR review: with no near-term plan to swap an
entity to a non-Pg source, no per-team ownership split, and new entity
types likely to be added over time, the boilerplate cost of separate
traits per entity (5 trait files, 5 adapter files, 5 generic params on
the orchestrator) outweighs the flexibility it bought.

Collapses to a single trait + single adapter:

  domain/ports.rs:
    BackfillSource — one trait with fetch_calls / fetch_chats /
    fetch_channels / fetch_documents / fetch_emails methods.

  outbound/source.rs:
    PgBackfillSource — one struct holding the DB pool + per-entity page
    sizes, implementing all five methods.

  domain/service.rs:
    BackfillOrchestrator<S, P> — two generics (down from six).
    BackfillServiceImpl alias collapses accordingly.

Adding a 6th entity is now: one new method on the trait + one new method
on PgBackfillSource + one new route + request type. No new generic
params, no new files. (Adding it later under the separate-trait shape
would have been more ceremony.)

Outbound directory shrinks:
  outbound/backfill/{calls,chats,channels,documents,emails}.rs   removed
  outbound/source.rs                                              added

Tests: the FakeSource in service tests no longer impls a trait at all —
it's a plain struct with a fetch_page method. drain_source takes a
closure rather than a trait object, so the test fakes don't have to
stub four unrelated methods. 17/17 passing.
