--- rust/cloud-storage/search_processing_service/src/outbound/backfill/mod.rs
@@ -1,5 +1,0 @@
-pub mod calls;
-pub mod channels;
-pub mod chats;
-pub mod documents;
-pub mod emails;
