+++ rust/cloud-storage/search_processing_service/src/outbound/mod.rs
@@ -1,2 +1,2 @@
-pub mod backfill;
 pub mod publisher;
+pub mod source;
